Solution1
