+++ src/components/Type/type.ts
@@ -1,0 +1,14 @@
+export type User = {
+  name: string;
+  id: number;
+  username: string;
+  email: string;
+};
+
+export type TodoElement = {
+  id: number;
+  title: string;
+  userId: number;
+  completed: boolean;
+  user: User;
+};
